Improve error messages when connections.toml does not exist (#28)
Previously when a named connection was requested but the
`connections.toml` file did not exist, the error message suggested
defining a section in the non-existent file without acknowledging it was
missing. This commit adds a more useful error message in this case.

Additionally, error messages for default connections now correctly refer
to the `[default]` section instead of an empty `[]` section.

Fixes #23.

Signed-off-by: Aaron Jacobs <aaron.jacobs@posit.co>