Add support for 'externalbrowser' authentication (#24)
This commit adds the machinery to support the "externalbrowser"
authenticator, which is an interactive sign-in flow for users with
access to a local browser. It includes session caching, which is a nice
ergonomic improvement for repeated calls.

(The underlying protocol is completely bespoke to Snowflake so it took me
a little while to get it actually working.)

You can test it as follows:

    snowflake_connection(
      account = <account>
      user = <username>,
      authenticator = "externalbrowser"
    ) |> snowflake_credentials()

This should open a browser window to the account's SSO provider.

Unfortunately this is all highly interactive code, so there is not much
to unit test.

Signed-off-by: Aaron Jacobs <aaron.jacobs@posit.co>